Speedup gibson multitrial nonhisto (avg,min,max)
Required for vcell-solvers commit 748fa9d7
